% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_stats_sum.R
\name{acc_stats_sum}
\alias{acc_stats_sum}
\title{Calculate binary classification accuracy}
\usage{
acc_stats_sum(tp, fp, tn, fn)
}
\arguments{
\item{tp}{True positives}

\item{fp}{False positives}

\item{tn}{True negatives}

\item{fn}{False negatives}
}
\value{
Classification accuracy and the True Skill Statistic
}
\description{
Calculate binary classification accuracy
}
\keyword{internal}
