% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accStatsSum.R
\name{accStatsSum}
\alias{accStatsSum}
\title{Calculate binary classification accuracy}
\usage{
accStatsSum(tp, fp, tn, fn)
}
\arguments{
\item{tp}{True positives}

\item{fp}{False positives}

\item{tn}{True negatives}

\item{fn}{False negatives}
}
\value{
Classification accuracy and the True Skill Statistic
}
\description{
Calculate binary classification accuracy
}

