% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPprepair.R
\name{callPprepair}
\alias{callPprepair}
\title{Function to make system call to polygon cleaning program pprepair}
\usage{
callPprepair(dirnm, spdfinname, spdfoutname, crs = crs)
}
\arguments{
\item{dirnm}{directory where the shapefiles will go}

\item{spdfinname}{Name of ESRI shapefile that needs cleaning (no .shp ext)}

\item{spdfoutname:}{Name of pprepair output shapefile (no .shp ext)}
}
\value{
spdf of cleaned polygons, read in from temporary .shp written to disk
}
\description{
Function to make system call to polygon cleaning program pprepair
}

