% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_cases.R
\name{case2_accuracy}
\alias{case2_accuracy}
\title{Case 2 accuracy metric}
\usage{
case2_accuracy(grid.poly, user.polys, in.acc.wt, out.acc.wt, count.acc.wt,
  new.in.acc.wt, new.out.acc.wt, frag.acc.wt, edge.acc.wt, cate.acc.wt)
}
\arguments{
\item{grid.poly}{Polygon of sampling grid}

\item{user.polys}{User's field boundary polygons}

\item{in.acc.wt}{Weighting for in grid map discrepancy}

\item{out.acc.wt}{Weighting for out of grid map discrepancy}

\item{count.acc.wt}{Weighting given to field count accuracy}

\item{new.in.acc.wt}{Weighting for in grid map in new score}

\item{new.out.acc.wt}{Weighting for out of grid map in new score}

\item{frag.acc.wt}{Weighting for fragmentation accuracy}

\item{edge.acc.wt}{Weighting for edge accuracy}
}
\description{
Case 2 accuracy metric
}
\details{
Accuracy assessment for case when worker maps fields but none exist
}
\keyword{internal}
