% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_cases.R
\name{case3_accuracy}
\alias{case3_accuracy}
\title{Case 3 accuracy metric}
\usage{
case3_accuracy(grid.poly, qaqc.polys, in.acc.wt, out.acc.wt, count.acc.wt,
  new.in.acc.wt, new.out.acc.wt, frag.acc.wt, edge.acc.wt, cate.acc.wt,
  acc.switch = 1)
}
\arguments{
\item{grid.poly}{Polygon of sampling grid}

\item{qaqc.polys}{QAQC polygons}

\item{in.acc.wt}{Weighting for in grid map discrepancy}

\item{out.acc.wt}{Weighting for out of grid map discrepancy}

\item{count.acc.wt}{Weighting given to field count accuracy}

\item{new.in.acc.wt}{Weighting for in grid map in new score}

\item{new.out.acc.wt}{Weighting for out of grid map in new score}

\item{frag.acc.wt}{Weighting for fragmentation accuracy}

\item{edge.acc.wt}{Weighting for edge accuracy}

\item{acc.switch}{1 for conventional accuracy, 2 for TSS}
}
\description{
Case 3 accuracy metric
}
\details{
Accuracy assessment for case when worker doesn't map fields but 
they do exist. Note that the TSS version of accuracy is still retainined 
here, but is no longer used because if return NULL values in certain cases.
}
\keyword{internal}
