% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_cases.R
\name{case4_accuracy}
\alias{case4_accuracy}
\title{Case 4 accuracy metric}
\usage{
case4_accuracy(grid.poly, user.polys, qaqc.polys, count.acc.wt, in.acc.wt,
  out.acc.wt, new.in.acc.wt, new.out.acc.wt, frag.acc.wt, edge.acc.wt,
  cate.acc.wt, edge.buf, cate.code, comments = "F", acc.switch = 1)
}
\arguments{
\item{grid.poly}{Polygon of sampling grid}

\item{user.polys}{User's field boundary polygons}

\item{qaqc.polys}{QAQC polygons}

\item{count.acc.wt}{Weighting given to field count accuracy}

\item{in.acc.wt}{Weighting for in grid map discrepancy}

\item{out.acc.wt}{Weighting for out of grid map discrepancy}

\item{new.in.acc.wt}{Weighting for in grid map in new score}

\item{new.out.acc.wt}{Weighting for out of grid map in new score}

\item{frag.acc.wt}{Weighting for fragmentation accuracy}

\item{edge.acc.wt}{Weighting for edge accuracy}

\item{edge.buf}{buffer for edge accuracy}

\item{cate.code}{the category codes reads from database}

\item{comments}{Should comments be printed, "T" or "F" (default)?}

\item{acc.switch}{1 for conventional accuracy, 2 for TSS}
}
\description{
Case 4 accuracy metric
}
\details{
Accuracy assessment for case when worker maps fields where they  
they do exist. Note that the TSS version of accuracy is still retainined 
here, but is no longer used because if return NULL values in certain cases.
}
\keyword{internal}
