% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_accuracy.R
\name{categorical_accuracy}
\alias{categorical_accuracy}
\title{Calculate categorical accuracy}
\usage{
categorical_accuracy(qaqc.polys, user.polys, cate.code)
}
\arguments{
\item{qaqc.polys}{qaqc polygons}

\item{user.polys}{user polygons (excluding polygons that users are unsure 
about field or no field)}

\item{cate.code}{the category codes reads from database}
}
\value{
categorical accuracy
}
\description{
Calculate categorical accuracy
}
\details{
categorical accuracy = field area with correct category labels 
divided by field area with correct field/no-field labels
}
