% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_cases.R
\name{checkexists}
\alias{checkexists}
\title{Function to handle missing objects or empty geometry sf objects & make NULL}
\usage{
checkexists(x, env)
}
\arguments{
\item{x}{A string naming the map object to check}

\item{env}{Environment variable}
}
\description{
Function to handle missing objects or empty geometry sf objects & make NULL
}
\details{
Used internally by case*_accuracy to check if objects exists, have 
geometries, and convert to NULL if they don't
}
\keyword{internal}
