% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanTempPolyFromWKt.R
\name{cleanTempPolyFromWKT}
\alias{cleanTempPolyFromWKT}
\title{Reads in PostGIS geometry and creates a temporary shapefile out}
\usage{
cleanTempPolyFromWKT(geom.tab, crs)
}
\arguments{
\item{geom.tab}{Dataframe with geometry (1st col) and identifiers (2nd col)}

\item{crs}{Coordinate reference system}
}
\value{
A cleaned SpatialPolygonsDataFrame
}
\description{
Reads in PostGIS geometry and creates a temporary shapefile out
}
\details{
Uses callPprepair function to clean up read in polygons and write
}

