% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countError.R
\name{countError}
\alias{countError}
\title{Calculates percent agreement between number of fields in qaqc and user kmls}
\usage{
countError(qaqc_rows, user_rows)
}
\arguments{
\item{qaqc_rows}{vector of count of QAQC rows, or NULL if one doesn't exist}

\item{kml}{User mapped fields, or NULL if they don't exist}
}
\value{
Score between 0-1
}
\description{
Calculates percent agreement between number of fields in qaqc and user kmls
}
\note{
Rearranges numerator and denominator of equation according to whether 
user mapped fields are more or less than QAQC fields
}
