% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_col_functions.R
\name{dummy_raster}
\alias{dummy_raster}
\title{Create a dummy raster that matches the dimensions of master grid}
\usage{
dummy_raster(xmin = -17.541, xmax = 51.419, ymin = -34.845,
  ymax = 37.54, res = 0.005)
}
\arguments{
\item{xmin}{Minimum X coordinate, defaulted to -17.541 of Africa grid}

\item{xmax}{Maximum X coordinate, defaulted to 51.419 of Africa grid}

\item{ymin}{Minimum Y coordinate, defaulted to -34.845 of Africa grid}

\item{ymax}{Maximum Y coordinate, defaulted to 37.540 of Africa grid}

\item{res}{Resolution of grid, default 0.005}
}
\value{
Unpopulated raster set to provided extent and resolution
}
\description{
Create a dummy raster that matches the dimensions of master grid
}
\details{
Set up dummy raster. Assumes EPSG 4326.
}
\keyword{internal}
