% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_maps.R
\name{error_maps}
\alias{error_maps}
\title{Plotting function for KMLAccuracy}
\usage{
error_maps(grid_poly, qaqc_poly, user_poly, inres, err_out, user_poly_out,
  qaqc_poly_out, tpout, fnout, pngout = TRUE)
}
\arguments{
\item{grid_poly}{sf polygon object of sampling grid}

\item{qaqc_poly}{sf polygon object of q polygon (assuming it exists)}

\item{user_poly}{sf polygon object of user maps (assuming it exists)}

\item{inres}{Output list from mapError}

\item{err_out}{Vector of error terms}

\item{user_poly_out}{sf polygon for portion of user map outside of grid}

\item{qaqc_poly_out}{sf polygon for portion of q map outside of grid}

\item{tpo}{sf polygon of correct user maps outside of grid (if exists)}

\item{fno}{sf polygon of false negative area outside of grid (if exists)}
}
\description{
Plotting function for KMLAccuracy
}
\details{
Not currently functional, but intended to provide replacement for plotting
code in KMLAccuracy
}
