% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_accuracy.R
\name{geometric_accuracy}
\alias{geometric_accuracy}
\title{Calculate fragmentation accuracy and edge accuracy}
\usage{
geometric_accuracy(qaqc.polys, user.polys, buf)
}
\arguments{
\item{qaqc.polys}{qaqc polygons}

\item{user.polys}{user-map polygons}

\item{buf}{buffer width for edge accuracy}
}
\description{
Calculate fragmentation accuracy and edge accuracy
}
\details{
This function targets on how the geometry of qaqc polygons are 
mapped by user polygons, which should be complementary to the in-grid and 
outgrid thematic accuracy.
}
\keyword{internal}
