% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDBName.R
\name{getDBName}
\alias{getDBName}
\title{Find correct root path and database name}
\usage{
getDBName(db.sandbox.name = "SouthAfricaSandbox",
  db.production.name = "SouthAfrica")
}
\arguments{
\item{db.sandbox.name}{Name of development server/database}

\item{db.production.name}{Name of production server/database}
}
\value{
Root path and database name in named vector
}
\description{
Find correct root path and database name
}
\note{
The function arguments currently default to SouthAfrica*, so expect
these to change with project upgrades
}

