% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_accuracy.R
\name{kml_accuracy}
\alias{kml_accuracy}
\title{Main function for running worker accuracy assessment}
\usage{
kml_accuracy(mtype, diam, prjsrid, kmlid, assignmentid, tryid,
  count.acc.wt, in.acc.wt, out.acc.wt, new.in.acc.wt, new.out.acc.wt,
  frag.acc.wt, edge.acc.wt, cate.acc.wt, edge.buf, acc.switch, comments,
  write.acc.db, draw.maps, pngout = TRUE, test, test.root, host = NULL)
}
\arguments{
\item{mtype}{"qa" or "tr" for qaqc or training types}

\item{diam}{diameter for master grids}

\item{prjsrid}{EPSG identifier for projected coordinate system}

\item{count.acc.wt}{Weighting given to field count error}

\item{in.acc.wt}{Weighting for in grid map discrepancy}

\item{out.acc.wt}{Weighting for out of grid map discrepancy}

\item{new.in.acc.wt}{Weighting for in grid map in new score}

\item{new.out.acc.wt}{Weighting for out of grid map in new score}

\item{frag.acc.wt}{Weighting for fragmentation accuracy}

\item{edge.acc.wt}{Weighting for edge accuracy}

\item{edge.buf}{buffer for edge accuracy}

\item{acc.switch}{in grid error metric: 1 = overall accuracy; 2 = TSS}

\item{comments}{For testing, can turn off (F) or on (T) print statements}

\item{write.acc.db}{Write error metrics into accuracy_data table ("T" or "F")}

\item{draw.maps}{Draw maps showing output error components ("T" or "F")}

\item{pngout}{Write maps to png file, TRUE (default) or FALSE (to screen)}

\item{test}{"Y" or "N" for offline testing mode (see Details)}

\item{host}{NULL or "crowdmapper.org", if testing from remote location}
}
\description{
Main function for running worker accuracy assessment
}
\details{
For the test argument, it can be set to "Y" if one wants to test 
only a single kmlid. In this case, the function code will pull the 
entire assignment_data and hit_data tables from the database to find the 
right assignment ids to test. This option must be set to "N" when in  
production. test.root allows one to simply the run the function to see if it 
is located in the correct working environment.
}
