% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMLAccuracy.R
\name{KMLAccuracy}
\alias{KMLAccuracy}
\title{Main function for running worker accuracy assessment}
\usage{
KMLAccuracy(mtype, kmlid, assignmentid, tryid, prjsrid, count.err.wt, in.err.wt,
  out.err.wt, err.switch, comments, write.err.db, draw.maps, test, test.root,
  user, password)
}
\arguments{
\item{prjsrid}{EPSG identifier for projected coordinate system}

\item{count.err.wt}{Weighting given to field count error}

\item{in.err.wt}{Weighting for in grid map discrepancy}

\item{out.err.wt}{Weighting for out of grid map discrepancy}

\item{err.switch}{in grid error metric: 1 = overall accuracy; 2 = TSS}

\item{comments}{For testing, can turn off (F) or on (T) print statements}

\item{write.err.db}{Write error metrics into error_data table ("T" or "F")}

\item{draw.maps}{Draw maps showing output error components ("T" or "F")}

\item{test}{"Y" or "N" for offline testing mode (see Details)}

\item{test.root}{"Y" or "N" for testing current working location}

\item{user}{User name for database connection}

\item{password}{Password for database connection}
}
\description{
Main function for running worker accuracy assessment
}
\details{
For the test argument, it can be set to "Y" if one wants to test a 
given only a single kmlid. In this case, the function code will pull the 
entire assignment_data and hit_data tables from the database to find the 
right assignment ids to test. This option must be set to "N" when in  
production. test.root allows one to simply the run the function to see if it 
is located in the correct working environment.
}

