% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_accuracy.R
\name{map_accuracy}
\alias{map_accuracy}
\title{Calculates mapping accuracy for polygons relative to a "true" set of polygons}
\usage{
map_accuracy(maps, truth, region)
}
\arguments{
\item{maps}{Polygons to assess}

\item{truth}{Polygons against which which accuracy will be assessed}

\item{region:}{A polygon defining the region in which accuracy is assessed}
}
\value{
Accuracy measures from user maps
}
\description{
Calculates mapping accuracy for polygons relative to a "true" set of polygons
}
\keyword{internal}
