% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{mapper_connect}
\alias{mapper_connect}
\title{Find correct root path and database name}
\usage{
mapper_connect(host = NULL)
}
\arguments{
\item{host}{NULL is running on crowdmapper, else crowdmapper.org for remote}
}
\value{
Database connection, root path, and database name in list
}
\description{
Find correct root path and database name
}
\note{
The function arguments currently default to SouthAfrica*, so expect 
these to change with project upgrades
}
