% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_to_gridpoly.R
\name{point_to_gridpoly}
\alias{point_to_gridpoly}
\title{Creates polygon grid box from point}
\usage{
point_to_gridpoly(xy, w, OldCRSobj, NewCRSobj)
}
\arguments{
\item{xy}{A data.frame containing x and y coordinates and name of grid point}

\item{w}{A value in meters specifiying 1/2 the diameter of the grid polygon}

\item{OldCRSobj}{The crs for the input polygon}

\item{NewCRSobj}{crs to transform polygon to}
}
\description{
Creates polygon grid box from point
}
\details{
This function is created for use by KMLgenerate, which uses it to 
convert grid points to polygons.
}
