% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyFromWkt.R
\name{polyFromWkt}
\alias{polyFromWkt}
\title{Creates a spatialPolygonsDataFrame from a geometry table read in from Postgis}
\usage{
polyFromWkt(geom.tab, crs)
}
\arguments{
\item{geom.tab}{Dataframe with geometry (1st col) and identifiers (2nd col)}

\item{crs}{Coordinate reference system}
}
\value{
A SpatialPolygonsDataFrame
}
\description{
Creates a spatialPolygonsDataFrame from a geometry table read in from Postgis
}
