% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_col_functions.R
\name{rowcol_from_xy}
\alias{rowcol_from_xy}
\title{Retrieve raster row and column numbers for x, y coordinate and raster}
\usage{
rowcol_from_xy(x, y, r, offset = -1)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}

\item{r}{Reference raster}

\item{offset}{Adjust to match referencing convention, e.g. 0-based (-1)}
}
\value{
Matrix of row and column number
}
\description{
Retrieve raster row and column numbers for x, y coordinate and raster
}
\details{
Used to find row and cell number for a given raster.
}
\examples{
library(rmapaccuracy)
r <- dummy_rast()
rowcol_from_xy(r, x = 8.9065, y = 37.5375)
}
\keyword{export}
