% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_upload.R
\name{s3_upload}
\alias{s3_upload}
\title{Upload spatial objects to S3 bucket}
\usage{
s3_upload(proj.root, bucketname, local.object, s3.dst, s3.filename)
}
\arguments{
\item{proj.root}{}

\item{bucketname}{the name of S3 bucket}

\item{local.object}{local spatial objects, and it has to be RasterLayer or sf
objects}

\item{s3.dst}{the folder directory of s3 bucket for putting local objects}

\item{s3.filename}{the name of local.object saved in s3}
}
\description{
Upload spatial objects to S3 bucket
}
