% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model_averaging.R
\name{bayes_model_averaging}
\alias{bayes_model_averaging}
\title{Bayesian Model Averaging P(theta|D) = ∑ weight * mapper posterior probability}
\usage{
bayes_model_averaging(bayes.polys, rasterextent, threshold)
}
\arguments{
\item{rasterextent}{the extent for the output}

\item{threshold}{the threshold for P(theta|D) to determine the label of pixels
as field or no field}

\item{bayes.poly}{a sf object has five columns: 
(1)posterior.field a are mapper posterior probability, meaning that
mappers' opinion for the possibility of field (we set 1 for sure category, 
and 0.5 for unsure);  
(3)max.field.lklh, (4) max.nofield.lklh, namely the producer's accuracy, which means
that given its label as field or no field, the maximum likelihood to be the 
mapper i; (5) score
Weight = max.nofield.lklh (or max.field.lklh) * score}
}
\value{
A list of three rasters--heat map, risk map, and consensus map
}
\description{
Core codes for Bayesian Model Averaging 
P(theta|D) = ∑ weight * mapper posterior probability
}
