% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTempPolyFromWKt.R
\name{cleanTempPolyFromWKT}
\alias{cleanTempPolyFromWKT}
\title{Reads in PostGIS geometry and creates a temporary shapefile out}
\usage{
cleanTempPolyFromWKT(unfixedsfc, crs)
}
\arguments{
\item{unfixedsfc}{sfc object for unfixed polygon}

\item{crs}{Coordinate reference system}
}
\value{
A cleaned sf object
}
\description{
Reads in PostGIS geometry and creates a temporary shapefile out
}
\details{
Uses callPprepair function to clean up read in polygons and write
}
\note{
Deprecated in current version of platform, which uses ST_MakeValid not 
pprepair
}
