% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_creation.R
\name{consensus_map_creation}
\alias{consensus_map_creation}
\title{Control the output of label maps, risk maps, and heat maps}
\usage{
consensus_map_creation(kmlid, kml.usage, mode, output.riskmap, diam, host,
  qsite = FALSE)
}
\arguments{
\item{kmlid}{}

\item{kml.usage}{the use of kml, could be 'train, 'validate' or 'holdout';
This parameter will determine the directory of S3 folder to store consensus 
maps.}

\item{mode}{the method for generating consensus: 'consensus', 'high' or 'low'.
'high' or 'low' means that using the highest or the lowest score worker maps}

\item{host}{NULL or "crowdmapper.org", if testing from remote location}

\item{qsite}{Q ir F site?  Default is FALSE}

\item{riskthres}{the threshold to select 'risk' pixels}
}
\value{
Sticks conflict/risk percentage pixels into database (kml_data) and
writes rasterized labels to S3 bucket.
}
\description{
Control the output of label maps, risk maps, and heat maps
}
