% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{get_db_name}
\alias{get_db_name}
\title{Find correct root path and database name}
\usage{
get_db_name()
}
\value{
Root path and database name in named vector
}
\description{
Find correct root path and database name
}
\note{
The function arguments currently default to Africa*, so expect 
these to change with project upgrades
}
