% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{getDBName}
\alias{getDBName}
\title{Find correct root path and database name}
\usage{
getDBName(db.sandbox.name = "AfricaSandbox", db.production.name = "Africa",
  db.tester.name = NULL, alt.root = NULL)
}
\arguments{
\item{db.sandbox.name}{Name of development server/database}

\item{db.production.name}{Name of production server/database}

\item{db.tester.name}{Name of non-sandbox user, for testing}

\item{alt.root}{An alternative root, for testing off of mapper VM}
}
\value{
Root path and database name in named vector
}
\description{
Find correct root path and database name
}
\note{
The function arguments currently default to Africa*, so expect 
these to change with project upgrades
}
