% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{mapper_connect}
\alias{mapper_connect}
\title{Find correct root path and database name}
\usage{
mapper_connect(host = NULL, user = NULL)
}
\arguments{
\item{host}{NULL is running on crowdmapper, else crowdmapper.org for remote}

\item{user}{NULL (effective user) or override with manually supplied name}
}
\value{
Database connection, root path, and database name in list
}
\description{
Find correct root path and database name
}
\details{
If you want to log into, say, the mapper user on a running instance
from a local machine, set the user argument to "mapper".
}
\note{
The function arguments currently default to SouthAfrica*, so expect 
these to change with project upgrades.
}
