% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_col_functions.R
\name{rowcol_from_xy}
\alias{rowcol_from_xy}
\title{Retrieve raster row and column numbers for x, y coordinate and raster}
\usage{
rowcol_from_xy(x, y, xmin = -17.541, xmax = 51.419, ymin = -34.845,
  ymax = 37.54, res = 0.005, offset = -1)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}

\item{xmin}{Minimum X coordinate, defaulted to -17.541 of Africa grid}

\item{xmax}{Maximum X coordinate, defaulted to 51.419 of Africa grid}

\item{ymin}{Minimum Y coordinate, defaulted to -34.845 of Africa grid}

\item{ymax}{Maximum Y coordinate, defaulted to 37.540 of Africa grid}

\item{res}{Resolution of grid, default 0.005}

\item{offset}{Adjust to match referencing convention, e.g. 0-based (-1)}
}
\value{
Matrix of row and column number
}
\description{
Retrieve raster row and column numbers for x, y coordinate and raster
}
\details{
Used to find row and cell number for a given raster.
}
\keyword{export}
